+++ WebContent/WEB-INF/views/admin/createSession.jsp
@@ -1,0 +1,34 @@
+<%@ taglib uri="http://tiles.apache.org/tags-tiles" prefix="tiles"%>  
+<%@ taglib prefix="form" uri="http://www.springframework.org/tags/form"%>
+<%@ taglib prefix="spring" uri="http://www.springframework.org/tags" %>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+
+<h1>Create new Session</h1>
+
+<form method="post" action="createSession">
+<label for="dates">Exam Session's starting: </label>
+<input id="startingDate" name="startingDate" type="date" value="01-01-2017"/>
+<br>
+<label for="dates">Exam Session's ending: </label>
+<input id="endingDate" name="endingDate" type="date" value="01-01-2017"/>
+
+<br>
+<label>DegreeCOURSE: </label>
+<select name="degreeCourse">
+	<option>Choose</option>
+	<c:forEach items="${degreeCourses}" var="degree">
+    	
+    	<option> <c:out value="${degree.name}"/></option>
+  	</c:forEach>
+  	</select>
+  	
+<br>
+<label>Academic Year: </label>
+<select name="academicYear">
+    <option value="Choose">Choose</option>
+    <option value="2016/2017">2016/2017</option>
+    <option value="2017/2018">2017/2018</option>
+  </select>
+  
+  <input type="submit" value="Submit">
+</form>
