+++ src/main/java/it/unical/asde/uam/controllers/ProfessorController.java
@@ -102,45 +102,6 @@ public class ProfessorController extends BaseController {
         return "professor/register";
     }
     
-    
-    
-    
-    @RequestMapping(value = "createSession", method = RequestMethod.GET)
-    public String openCreateNewSession(Model model) {
-    	DegreeCourseDAO degreeCourseDao = (DegreeCourseDAO) context.getBean("degreeCourseDAO");
-    	ArrayList<DegreeCourse> allDegrees = (ArrayList<DegreeCourse>) degreeCourseDao.getAllDegrees();
-    	model.addAttribute("degreeCourses", allDegrees);
-    	
-        return "professor/createSession";
-    }
-    
-    @RequestMapping(value = "createSession", method = RequestMethod.POST)
-    public String createNewSession(@RequestParam("startingDate") String startingDateString, @RequestParam("endingDate") String endingDateString,
-    		@RequestParam("degreeCourse") String degreeCourseName, @RequestParam("academicYear") String academicYear,
-    		HttpServletRequest request) throws ParseException{
-    	
-    	ProfessorDAO professorDao = (ProfessorDAO) context.getBean("professorDAO");
-    	if(professorDao.checkExamSession(request.getParameter("startingDate"),request.getParameter("endingDate"), request.getParameter("academicYear")))
-    	{
-    		DegreeCourseDAO degreeCourseDAO = (DegreeCourseDAO) context.getBean("degreeCourseDAO");
-    		DegreeCourse degreeCourse = degreeCourseDAO.getDegreeCourseByName(degreeCourseName);
-    		
-    		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-    		Date startingDate = null;
-   			startingDate = sdf.parse(startingDateString);
-   			Date endingDate = null;
-   			endingDate = sdf.parse(endingDateString);
-    		
-   			ExamSessionDAO examSessionDao = (ExamSessionDAO) context.getBean("examSessionDAO");
-    		ExamSession examSession = new ExamSession(startingDate, endingDate, academicYear, degreeCourse);
-
-    		examSessionDao.create(examSession);
-    		ArrayList<ExamSession> allExamSessions = (ArrayList<ExamSession>) examSessionDao.getAllExamSession();//professorDao.listAllSession();
-    		return "professor/dashboard";
-    	}
-    	else return "professor/errorExamSession";
-    }
-    
     @RequestMapping(value ="viewAllSession", method = RequestMethod.GET)
     public String viewAllSession(Model model){
     	ExamSessionDAO examSessionDao = (ExamSessionDAO) context.getBean("examSessionDAO");
