+++ src/main/java/it/unical/asde/uam/persistence/ProfessorDAO.java
@@ -35,8 +35,6 @@ public interface ProfessorDAO {
     
     boolean register(Professor u);
 
-    boolean checkExamSession(String startingDate, String endingDate, String academicYear);
-    
     ArrayList<ExamSession> listAllSession();
     
     ArrayList<Professor> getAllProfessor();
